Factor out additive and binary
